third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/maxPool2d.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] maxPool2d float32 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] maxPool2d float32 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual 99.28312683105469 should be close enough to expected -39.03501892089844 by ULP distance: expected a number less than or equal to 0n but got 2229449938n
[FAIL] maxPool2d float32 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float32 actual 99.28312683105469 should be close enough to expected -39.03501892089844 by ULP distance: expected a number less than or equal to 0n but got 2229449938n
[FAIL] maxPool2d float16 4D constant tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual -0.0078125 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 18891
[FAIL] maxPool2d float16 4D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.windowDimensions
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.padding
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.strides
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.dilations
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': Pool2d in tflite doesn't support dilations."
[FAIL] maxPool2d float16 4D tensor options.layout=nchw
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.layout=nhwc
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.outputSizes ignores options.roundingType=floor
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.outputSizes ignores options.roundingType=ceil
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 99.3125 by ULP distance: expected a number less than or equal to 0 but got 22069
[FAIL] maxPool2d float16 4D tensor options.dilations with options.strides
  assert_less_than_equal: assert_array_approx_equals_ulp: test maxPool2d float16 actual 0 should be close enough to expected 60.90625 by ULP distance: expected a number less than or equal to 0 but got 21405
Harness: the test ran to completion.
